requirements.txt
python-dotenv
streamlit
sqlalchemy
pymysql